Actualización de pantalla de registro y eliminación de pantalla de verificación de correo electrónico, Creado Ventana Emergente para verificación de correro Electrónico
